fix: change branch name in sparseCheckout function (#116)
### Motivation
the `sparseCheckout` function is currently only used in the aws-auth helper to pull a file from the terraform repo. Since the writing of this function, the main branch name was changed from `master` to `main`, which caused this function to error silently. 

### Changes
change the branch name from `master` to `main` to checkout